Avoid accumulating "empty" nodes
This commit includes tests which show that adding entries to a trie
and deleting them all leaves "empty" nodes (i.e. those with a nil
item) around. In long-running processes, this leads to excessive
memory usage. The fix is to prune some unnecessary nodes during
deletion.

Note: there is a trade-off against performance. We attempt to prune a
node which has been deleted, but we don't attempt to prune its
ancestors. The latter approach made performance of the "random kitchen
sink" test worse by at least two orders of magnitude.

Signed-off-by: Glyn Normington <gnormington@pivotal.io>